Refactor AppSidebar into smaller files
Refactors the AppSidebar component into smaller, more manageable files. This improves code organization and maintainability.
