+++ src/components/Layout/Sidebar/index.ts
@@ -1,0 +1,2 @@
+
+export { AppSidebar } from './AppSidebar';
